+++ typeit/__init__.py
@@ -1,6 +1,7 @@
 from . import flags
 from . import sums
+from . import custom_types
 from .combinator.constructor import type_constructor, TypeConstructor
 from .schema.errors import Error
 
-__all__ = ('TypeConstructor', 'type_constructor', 'flags', 'sums', 'Error')
+__all__ = ('TypeConstructor', 'type_constructor', 'flags', 'sums', 'Error', 'custom_types')