+++ typeit/custom_types/__init__.py
@@ -1,0 +1,3 @@
+from .json_string import JsonString
+
+__all__ = ('JsonString', )